// frontend/src/types/supplier.ts
export interface Supplier {
    id: number;
    code: string;
    name: string;
    tax_id: string;
    address: string;
    contact_phone: string;
    type: string;
    delivery_schedule: string;
    temperature_requirements: string;
    sales_rep_name: string;
    sales_rep_id: string;
    created_at: string;
    updated_at: string;
}